+++ src/main/java/io/reactiverse/es4x/dynalink/ES4XJSONLinker.java
@@ -6,14 +6,15 @@ import jdk.dynalink.linker.*;
 import jdk.nashorn.api.scripting.ScriptUtils;
 
 import java.lang.invoke.*;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.function.Supplier;
 
 public class ES4XJSONLinker implements GuardingDynamicLinker, GuardingTypeConverterFactory {
 
   private static final MethodHandle TO_JSONOBJECT;
   private static final MethodHandle TO_JSONARRAY;
+
+  private final Map<Class, DataObjectLinker> linkers = new IdentityHashMap<>();
 
   static {
     final MethodHandles.Lookup lookup = MethodHandles.lookup();
@@ -48,6 +49,25 @@ public class ES4XJSONLinker implements GuardingDynamicLinker, GuardingTypeConver
     if (targetType.isAssignableFrom(JsonArray.class)) {
       return new GuardedInvocation(TO_JSONARRAY);
     }
+    // will attempt to cache the linker if not available yet
+    DataObjectLinker linker;
+
+    if (linkers.containsKey(targetType)) {
+      linker = linkers.get(targetType);
+      if (linker != null) {
+        return new GuardedInvocation(linker.getHandler());
+      }
+    } else {
+      // first time check
+      try {
+        linker = new DataObjectLinker(targetType);
+        linkers.put(targetType, linker);
+        return new GuardedInvocation(linker.getHandler());
+      } catch (NoSuchMethodException | IllegalAccessException e) {
+        linkers.put(targetType, null);
+      }
+    }
+
     // we can't convert
     return null;
   }